pyproject.toml
[tool.poetry]
name = "organization-catalog"
version = "0.1.0"
description = "Тестовое задание для Secunda"
authors = ["Rav <rrrdobr@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
geoalchemy2 = "0.14.3"
asyncpg = "^0.29.0"
pydantic = "^2.5.1"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
shapely = "2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
httpx = "^0.25.1"
pytest-asyncio = "^0.23.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
